fixing another service fixing covers, separated verification tests to new test class
